+++ extensions/shared/library/src/main/java/app/revanced/extension/shared/Utils.java
@@ -9,6 +9,7 @@ import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.content.res.Configuration;
 import android.content.res.Resources;
+import android.graphics.Color;
 import android.net.ConnectivityManager;
 import android.os.Build;
 import android.os.Bundle;
@@ -798,5 +799,15 @@ public class Utils {
         if (editTextDialogStyle != 0) {
             builder.getContext().setTheme(editTextDialogStyle);
         }
+    }
+
+    /**
+     * Parse a color resource or hex code to an int representation of the color.
+     */
+    public static int getColorFromString(String colorString) throws IllegalArgumentException, Resources.NotFoundException {
+        if (colorString.startsWith("#")) {
+            return Color.parseColor(colorString);
+        }
+        return getResourceColor(colorString);
     }
 }