+++ extensions/youtube/src/main/java/app/revanced/extension/youtube/patches/theme/SeekbarColorPatch.java
@@ -226,6 +226,7 @@ public final class SeekbarColorPatch {
     }
 
     private static String loadRawResourceAsString(int resourceId) {
+        //noinspection CharsetObjectCanBeUsed
         try (InputStream inputStream = Utils.getContext().getResources().openRawResource(resourceId);
              Scanner scanner = new Scanner(inputStream, StandardCharsets.UTF_8.name()).useDelimiter("\\A")) {
             return scanner.next();
@@ -281,6 +282,20 @@ public final class SeekbarColorPatch {
 
     /**
      * Injection point.
+     * 19.49+
+     */
+    public static int[] getPlayerLinearGradient(int[] original, int x0, int y1) {
+        // This hook is used for both the player and the feed.
+        // Feed usage always has x0 and y1 value of zero, and the player is always non zero.
+        if (HIDE_SEEKBAR_THUMBNAIL_ENABLED && x0 == 0 && y1 == 0) {
+            return HIDDEN_SEEKBAR_GRADIENT_COLORS;
+        }
+        return getPlayerLinearGradient(original);
+    }
+
+    /**
+     * Injection point.
+     * Pre 19.49
      */
     public static int[] getPlayerLinearGradient(int[] original) {
         return SEEKBAR_CUSTOM_COLOR_ENABLED