+++ patches/src/main/kotlin/app/revanced/patches/youtube/layout/hide/general/HideLayoutComponentsPatch.kt
@@ -20,11 +20,10 @@ import app.revanced.patches.shared.misc.settings.preference.*
 import app.revanced.patches.youtube.misc.litho.filter.addLithoFilter
 import app.revanced.patches.youtube.misc.litho.filter.lithoFilterPatch
 import app.revanced.patches.youtube.misc.navigation.navigationBarHookPatch
-import app.revanced.patches.youtube.misc.playservice.is_19_47_or_greater
 import app.revanced.patches.youtube.misc.playservice.is_20_07_or_greater
-import app.revanced.patches.youtube.misc.playservice.versionCheckPatch
 import app.revanced.patches.youtube.misc.settings.PreferenceScreen
 import app.revanced.patches.youtube.misc.settings.settingsPatch
+import app.revanced.util.findFreeRegister
 import app.revanced.util.findInstructionIndicesReversedOrThrow
 import app.revanced.util.getReference
 import app.revanced.util.indexOfFirstInstructionOrThrow
@@ -123,7 +122,6 @@ val hideLayoutComponentsPatch = bytecodePatch(
         addResourcesPatch,
         hideLayoutComponentsResourcePatch,
         navigationBarHookPatch,
-        versionCheckPatch
     )
 
     compatibleWith(
@@ -132,8 +130,6 @@ val hideLayoutComponentsPatch = bytecodePatch(
             "19.25.37",
             "19.34.42",
             "19.43.41",
-            "19.45.38",
-            "19.46.42",
             "19.47.53",
             "20.07.39",
         ),
@@ -254,17 +250,16 @@ val hideLayoutComponentsPatch = bytecodePatch(
         (if (is_20_07_or_greater) parseElementFromBufferFingerprint
         else parseElementFromBufferLegacyFingerprint).let {
             it.method.apply {
-                // Target code is a mess with a lot of register moves.
-                // There is no simple way to find a free register for all versions so this is hard coded.
-                val freeRegister = if (is_19_47_or_greater) 6 else 0
                 val byteArrayParameter = "p3"
                 val startIndex = it.patternMatch!!.startIndex
                 val conversionContextRegister = getInstruction<TwoRegisterInstruction>(startIndex).registerA
                 val returnEmptyComponentInstruction = instructions.last { it.opcode == Opcode.INVOKE_STATIC }
                 val returnEmptyComponentRegister = (returnEmptyComponentInstruction as FiveRegisterInstruction).registerC
+                val insertIndex = startIndex + 1
+                val freeRegister = findFreeRegister(insertIndex, conversionContextRegister, returnEmptyComponentRegister)
 
                 addInstructionsWithLabels(
-                    startIndex + 1,
+                    insertIndex,
                     """
                         invoke-static { v$conversionContextRegister, $byteArrayParameter }, $LAYOUT_COMPONENTS_FILTER_CLASS_DESCRIPTOR->filterMixPlaylists(Ljava/lang/Object;[B)Z
                         move-result v$freeRegister 
@@ -272,7 +267,7 @@ val hideLayoutComponentsPatch = bytecodePatch(
                         move-object v$returnEmptyComponentRegister, p1   # Required for 19.47
                         goto :return_empty_component
                         :show
-                        const/4 v$freeRegister, 0x0   # Restore register, required for 19.16
+                        nop
                     """,
                     ExternalLabel("return_empty_component", returnEmptyComponentInstruction),
                 )